pyproject.toml
[project]
name = "twicketsbot"
version = "0.1.0"
description = "A Discord ticketing bot"
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.10"

[tool.poetry.dependencies]
python = ">=3.10"
"discord.py" = "^2.0"
pyyaml = "^6.0"
aiosqlite = "^0.19"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
